Refactoring: package 'diagramexporters' renamed as 'source'.
