+++ plugins/hu.elte.txtuml.layout.export/src/hu/elte/txtuml/layout/export/impl/ElementExporterImpl.java
@@ -3,8 +3,6 @@ package hu.elte.txtuml.layout.export.impl;
 import hu.elte.txtuml.api.layout.Alignment;
 import hu.elte.txtuml.api.layout.Contains;
 import hu.elte.txtuml.layout.export.DiagramType;
-import hu.elte.txtuml.layout.export.diagramexporters.ModelId;
-import hu.elte.txtuml.layout.export.diagramexporters.SourceExporter;
 import hu.elte.txtuml.layout.export.elementinfo.ConcreteElementInfo;
 import hu.elte.txtuml.layout.export.elementinfo.ElementInfo;
 import hu.elte.txtuml.layout.export.elementinfo.GroupInfo;
@@ -22,6 +20,8 @@ import hu.elte.txtuml.layout.export.interfaces.NodeMap;
 import hu.elte.txtuml.layout.export.problems.ElementExportationException;
 import hu.elte.txtuml.layout.export.problems.ProblemReporter;
 import hu.elte.txtuml.layout.export.problems.Utils;
+import hu.elte.txtuml.layout.export.source.ModelId;
+import hu.elte.txtuml.layout.export.source.SourceExporter;
 import hu.elte.txtuml.layout.visualizer.model.LineAssociation;
 import hu.elte.txtuml.layout.visualizer.model.RectangleObject;
 import hu.elte.txtuml.utils.Pair;
