+++ plugins/hu.elte.txtuml.layout.export/src/hu/elte/txtuml/layout/export/source/ModelId.java
@@ -1,4 +1,4 @@
-package hu.elte.txtuml.layout.export.diagramexporters;
+package hu.elte.txtuml.layout.export.source;
 
 public interface ModelId {
